Stop Building up Stack on Sequence Lower
Lowering the sequence no longer leave the previous values on the
stack. Each element in a sequence is "stashed" in local 0 and the
final one pushed back on the stack once the sequence is complete.

A better solution to this would be inserting a `pop` instruction in
between each element in the sequence.