Parse and Lower `if` Special Form
If forms have either a sinle `ifTrue` body or a pair of `ifTrue` and
`ifFalse` bodies. Handling of these forms begins at binding. A new
bound expression type is needed to handle this case. For lowering we
emit a pair of jumps and lower either a `null` or the else body
depending on if one is provided.

According to the spec if the else block is emitted the value returned
is `undefiend` so returning `null` in this case is acceptable. It may
be a better idea to follow other Scheme implementation's tracks and
provide a separate `undefined` value which we use in these cases
though. Just emitting `new object()` would be enough as a
first-cut. For the time being `null` is OK though.