Improve Eval Return Type
Eval now returns the un-coerced object that results from a given
expression rather than the exit code that it did before. This is
acchived by adding a new `$ScriptBody` method that the code is lowered
into rather than the `Main` method. The `Main` method now just deferrs
to this `$ScriptBody` and then runs the epilogue to coerce the return
type to an exit status.