Add `EmitCtx` to Flow State in Emit
Starting to break apart the compiler so it's a bit more composable. As
more functionality is added and new forms are supported we will need to
flow increasing amounts of information around in the `emit` phase. To
help with this a new `EmitCtx` type is added. New information can be
added to this context without having to ensure it is flowed correctly
through all of `emitExpression`'s recursion.