Improve Formatting of Lambdas in External Repr
The external representation of a lambda should _in theory_ be something
that can be read back in as a value and will produce the same object.
Switch to include the method name in the representation.

This is all a bit of a moot point right now as we don't support reading
in other symbol types yet.