Make Identifier Parsing R7RS Compiliant
Adds the example identifiers from the spec, and adjusts identifier
parsing to follow the semantics of R7RS. This mainly resolves around
identifiers made up of `.`, `-` and other 'numeric' non-digit
characters. The single `.` is not a symbol on its own which adds a
slight wrinkle.

We no longer use the `Identifier` parser from `FParsec` for this as
there isn't really any need. Scheme identifiers allow any characters
in any position and instead exclude identifiers which parse as a
number prefix. FParsec handles this case for us by the ordering of the
`choose` cases.

Choose ordering is also exploited to ensure that `.` is parsed
correctly iff it is not the beginning of a full identifier symbol.