FIX: Push `this` _before_ Other Args
This was the cause of the null reerence throwing. By pushing the
instance and the arguments in the wrong order we were trying to
interpret whatever happened to be the first argument as the env
object. Hence the reason things _appeared_ to work when there
wasn't an argument but failed otherwise.