Initialise Environment Eagerly
When emitting the body of a lambda emit the code to initialise and
hoist at the beginning of the body. This makes sure the environment is
_always_ available to write into.

Thanks to the re-write pass we can get away with ony hoisting arguments
into the environment. This means no dynamic hoisting is needed _at all_
only locals are moved to an environment.

To correctly perform the hoists up-front we now need the _full list_ of
environment slots in the bound tree rather than just the _size_ of the
environment.