Re-Enable Parser Tests
Porting over coverage from the old parse tests. We don't have parity
when it comes to the treatment of the `@` character. Old parser saw it
as an identifier. The grammar in R7RS seems to hint that this should be
treated as a malformed imaginary number `r@theta`.

Better testing of cooked identifiers will require a new layer over the
current untyped tree which performs the 'cooking'.