Functional Lexer Interface
Make the lexer type itself internal. Introduce a new functional
interface to the lexer. We still need the lexer interface _for now_ to
provide the parser with rich position information. Ideally the lexer
could be _only_ accessed via the functional interface. This would need
a bigger restructuring of parsing however.