Handle `expect` at EOF
Update the parser to allow bumpt to _optionally_ return a token. If
no token is retuend then we will just emit an empty token as the error
rather than throwing. This fixes the test case which fails to provide a
closing `)`, causing `expect` to be called when no tokens remain.