Improve CLI Option Naming
Add better names to most option values. This improves the help text a
little. It would be _super nice_ to be able to use proper sub-commands
to properly parse the disjoint operations supported by the compiler.
It doens't seem like that is quite possible right now with the default
behaviour as we can't use `MainCommand`.