Emit Hidden Sequnece Points for Missing Locations
When the location of a point is missing we can use a hidden sequence
point rahter than omitting the sequence point entirely. This ensures
that the debugger knows the sequence point exists but that the source
location is hidden from it.

Given that missing locations usually come from fabricated syntax as part
of macro expansion this seems like the best solution.