Offsets and Locations
Some work to improve the handling of text locations by the
`TextDocument`. This makes sure that locations are allways at a
consistent line colum offset. Previously the first line was 0-based and
the remaining lines 1-based.

Still not totally sure if we want the position objects to be 0-based
like Firethorn or 1-based ready for output to the screen.